refactor: 人口構成データの取得部分のコンポーネントの分離
